Moving file validations to the FileUploadable concern
