+++ app/assets/javascripts/bs-file-input-init.js
@@ -1,3 +1,3 @@
 $(document).ready(function () {
     bsCustomFileInput.init();
-})
+});
